Fixed VS10 warning

+++ Infovis/Testing/Cxx/TestRealDataPDescriptiveStatisticsMPI.cxx
@@ -538,7 +538,7 @@ int main( int argc, char** argv )
     }
 
   // Fill process dimensionality with ones if not provided or incomplete
-  unsigned int missingDim = 3 - procDim.size();
+  int missingDim = 3 - static_cast<int>( procDim.size() );
   for ( unsigned int d = 0; d < missingDim; ++ d )
     {
     procDim.push_back( 1 );
